Semana09 dia03 haciendo separacion de componentes
