+++ Dia01-github/semana09/dia02-03-react-todo-app/src/components/ToDoHeader.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+// export const ToDoHeader = (props) => {
+export const ToDoHeader = ({title}) => { //destructuracion de props
+  return (
+    <h1 className="text-2xl font-bold text-center">{title}</h1>
+  )
+}
+
+export default ToDoHeader
